refactor: moving plugin runner test to test package
