+++ test/plugin/test_helper.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package plugin
 
 import (
 	"github.com/apache/incubator-devlake/errors"
